big tidy up pass on godot_cfg
